use off_t instead of int for sizes (start of work on handling > 2GB files)

+++ src/decompress.h
@@ -13,8 +13,8 @@ typedef enum {
     AG_XZ,
 } ag_compression_type;
 
-ag_compression_type is_zipped(const void* buf, const int buf_len);
+ag_compression_type is_zipped(const void* buf, const off_t buf_len);
 
-void* decompress(const ag_compression_type zip_type, const void* buf, const int buf_len, const char* dir_full_path, int* new_buf_len);
+void* decompress(const ag_compression_type zip_type, const void* buf, const off_t buf_len, const char* dir_full_path, off_t* new_buf_len);
 #endif
