+++ src/search.h
@@ -60,7 +60,7 @@ typedef struct {
 
 symdir_t *symhash;
 
-void search_buf(const char *buf, const int buf_len,
+void search_buf(const char *buf, const off_t buf_len,
                 const char *dir_full_path);
 void search_stream(FILE *stream, const char *path);
 void search_file(const char *file_full_path);
